webapi/tct-webmessaging-w3c-tests/webmessaging/w3c/postMessage_Date.htm

<!DOCTYPE html>
<html>
<head>
<title> postMessage with Date object </title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>

<div style="display:none">
    <iframe width="70%" onload="PostMessageTest()" src="http://127.0.0.1:8081/opt/tct-webmessaging-w3c-tests/webmessaging/w3c/support/ChildWindowPostMessage.htm"></iframe>
</div>

<script>
    setup({timeout:500});
    
    var description = "Test Description: Messages can contain JavaScript values (e.g., Dates).";
    
    var t = async_test(description);
    
    var DATA = new Date();
    var TYPE = "object";
    var TARGET = document.querySelector("iframe");
                
    function PostMessageTest()
    {
        TARGET.contentWindow.postMessage(DATA, "*");
    }
    
    window.onmessage = t.step_func(function(e)
    {
        assert_equals(typeof(e.data), TYPE);
        assert_equals(e.data.valueOf(), DATA.valueOf());
        assert_not_equals(e.data, DATA, "Object is cloned");
        t.done();
    });
</script>
</body>
</html>
